Move the contents of the py3exiv2 directory to the top level.
The only thing that wasn’t in py3exiv2 was an older copy of setup.py
that doesn’t appear to have been used in quite some time, and could
not have worked with the directory tree organized the way it was.
That file is replaced with the setup.py that was inside py3exiv2/.

Also, further cleanup on .gitignore.