Fix all DeprecationWarnings and ResourceWarnings in the testsuite.
Mostly this is just converting `failUnless...` methods to the
equivalent `assert...` methods, with no semantic difference.

testutils.CheckFileSum is changed to use a with-block to
open the file, thus not relying on prompt garbage collection
to close the file again.

In two places in test/xmp.py, assertRaises was being called
incorrectly and the function under test was not actually being tested.
This was masking an actual bug in the function under
test (`XmpTag._convert_to_python`), which is corrected.