Whitespace cleanup.
Add a .gitattributes file to tell Git which files are binary, which
are expected to have tabs, and which are expected to have DOS line
endings.  Then fix all whitespace errors reported by

    $ git diff --check -r f3cd66e36099170580f4c11824771b3b5b90e688

(f3cd66... is the oldest revision in the repo) except for a spurious
report of a stray conflict marker in the README, which will be
addressed separately.

Except for the addition of .gitattributes, this commit only alters
white space.