Mechanical reformat of setup.py and src/, using YAPF.
All of the Python code outside the testsuite is now formatted to the
PEP-8 standard with some minor tweaks (see `.style.yapf` for details).
The changes only affect whitespace, except that in two places a trailing
comma is added to a long [ ... ] literal to trigger YAPF’s “put each
element of this list on its own line” heuristic.

(Thanks to the Trio project for turning me on to YAPF!)

(The testsuite will be handled separately, because there are unusual
constructs in some of those files that make YAPF barf.)