Introduce Python script to drive CI builds.
Since we may wind up having to download and build Boost.Python and/or
libexiv2 from source, and since we want to build on all three of
Linux, OSX, and Windows eventually, it’s going to wind up cleaner to
have a separate CI driver script in Python, rather than adding ever
more complicated shell scripts to azure-pipelines.yml.

The new script attempts to install the C-level build dependencies, so
it has a fighting chance of actually working.