Reorganize test/ so pytest auto-discovery works correctly.
pytest is pickier about the names of .py files containing tests than
setup.py test is.  It wants them all to be named test_something.py.
This can be overridden in setup.cfg, but it’s a good convention, it
makes a space for helper modules that don’t contain tests in test/.
Which we actually have one of (testutils.py), so.

For clarity’s sake, testutils.py is renamed to helpers.py.

setup.py test still works.