Use RAII to release and re-acquire the GIL.
This replaces a bunch of code that intercepts C++ exceptions by hand
and re-throws them after re-acquiring the GIL.  The “release_gil”
object’s destructor will automatically re-acquire the GIL both when it
goes out of scope normally, and when an exception is thrown out of its
block.

Image::getDataBuffer is not converted in this patch, because it
requires somewhat more work to be completely exception-safe, and
because it will make sense to remove its unnecessary double copy at
the same time.

Also add Emacs “local variables” annotations to the bottom of
_libexiv2.cpp so its auto-indenter will use the proper style for this
file.