Overhaul Image::getDataBuffer.
Use RAII in Image::getDataBuffer to handle both the Python GIL, and
saving and restoring the read position within the image file.  This
means that if libexiv2 throws an exception while reading from the
image, the file will be properly reset to its previous state.

Also, read directly into a freshly allocated Python bytes object,
which allows us to read in large chunks instead of byte-by-byte, and
eliminates an intermediate copy.