Replace Boost.Python with Cython.
This is a minimal conversion, which isn’t so much wrapping the Exiv2
library interfaces themselves, as it is wrapping the guts of the
previous Boost-dependent wrapper.  Further patches will clean this up
somewhat.  It’s still probably going to be clumsy in places because
Exiv2 uses C++ in ways that interact badly with Cython’s limitations,
but it’s worth it to get rid of the dependency on Boost.  (In
particular, building binary wheels in CI would have involved building
Boost from source for each supported Python version, which was
completely out of the question.)

This has minor consequences for the public API: error conditions may
not throw exactly the same exceptions that they used to, and byte
strings may appear in places where Unicode strings formerly did, or
vice versa.