azure-build.py: New action ‘build-and-test-sdist’.
This action is kinda like Automake’s “distcheck”: it runs the tests
in-place on the git checkout, it builds both a binary wheel and an
sdist tarball from the git checkout, and then it does everything all
over again _from_ the sdist tarball, and verifies that the test
results, the sdists, and the wheels are identical—byte-for-byte
identical sdists and wheels, equivalent test and coverage results.

In order to make this work we had to implement a bunch of
reproducibility hacks in both azure-build.py and setup.py.  Some of
these will be replaced with less hacky, er, hacks, Real Soon Now.
(In particular, we will not be hardwiring GCC’s -fdebug-prefix-map
option forever.)