Use pre-commit for per-changeset linting.
This handy tool can be found at <https://pre-commit.com/>; install
with pip or pipx and then run `pre-commit install` in your checkout
to activate it as a git pre-commit hook.  The same checks will also
be run on all changed files in CI.

Right now it does a bunch of whitespace-related checks, plus flake8
and YAML validation.  We might do type checking and/or autoformatting
in the future.