Overhaul testsuite using pytest features.
This is quite a bit easier to read, and pytest’s powerful
test-parametrization mechanism can be used to get better diagnostics.
(The number of tests increases from ~180 to ~650; this is because many
tests with internal loops over data tables have been converted to
parametrized tests.  There is no actual increase in the set of things
being tested.)

Also the testsuite is now less sloppy about temporary files.