(Attempt to) patch the libexiv2 testsuite for -DSUPPRESS_WARNINGS.
There are a bunch of tests in the libexiv2 testsuite that expect
stderr output from the library that is disabled by -DSUPPRESS_WARNINGS.
Patch it to expect no stderr output from the library.  This is the
simplest change that has a chance of working, but it could be too big
of a hammer, we’ll see.