More nuanced fixup of libexiv2 testsuite for -DSUPPRESS_WARNINGS.
Turns out not all of the expected stderr output is caused by
diagnostics that -DSUPPRESS_MESSAGES disables.  Also turns out some of
the old shell-script-based tests are expecting some of those messages,
too.  Programmatic edits embedded in azure-build.py are not going to
be practical for this; we need the full power of patch(1).
Accordingly, the existing Windows fixups are converted to patch files
as well.

Trailing whitespace is normal, and significant, in patch files;
exclude them from the trailing-whitespace pre-commit fixup.